Moved solution in last commit from getch.py to gaot_plus_plus.py.
Reverted changes to getch.py. Only use getch() if stdin is a tty. This fixes an exception that occurs when piping or redirecting input to the interpreter. gaot_plus_plus.py seems like a better place to put this logic since the purpose of the getch module is to provide unbuffered input from a terminal.

There is a small difference in the behavior of these two solutions when running a simple "cat" program with stdin redirected from a file. The former stops reading from the input file if an EOF byte (0x04) is encountered. The new solution stops reading if a null byte (0x00) is encountered. I don't think there is a solution for copying arbitrary binary files since Gaot++ doesn't provide a way to detect the true end of a file. (If stdin is a tty, then control-D still causes the cat program to exit).